+++ python/python/tests/test_async_client.py
@@ -77,6 +77,7 @@ from glide.config import (
     ServerCredentials,
 )
 from glide.constants import OK, TEncodable, TFunctionStatsResponse, TResult
+from glide.exceptions import TimeoutError as GlideTimeoutError
 from glide.glide_client import GlideClient, GlideClusterClient, TGlideClient
 from glide.routes import (
     AllNodes,
@@ -96,6 +97,7 @@ from tests.utils.utils import (
     compare_maps,
     convert_bytes_to_string_object,
     convert_string_to_bytes_object,
+    create_lua_lib_with_long_running_function,
     generate_lua_lib_code,
     get_first_result,
     get_random_string,
@@ -6816,6 +6818,123 @@ class TestCommands:
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_xinfo_stream(
+        self, glide_client: TGlideClient, cluster_mode, protocol
+    ):
+        key = get_random_string(10)
+        group_name = get_random_string(10)
+        consumer = get_random_string(10)
+        stream_id0_0 = "0-0"
+        stream_id1_0 = "1-0"
+        stream_id1_1 = "1-1"
+
+        # setup: add stream entry, create consumer group and consumer, read from stream with consumer
+        assert (
+            await glide_client.xadd(
+                key, [("a", "b"), ("c", "d")], StreamAddOptions(stream_id1_0)
+            )
+            == stream_id1_0.encode()
+        )
+        assert await glide_client.xgroup_create(key, group_name, stream_id0_0) == OK
+        assert await glide_client.xreadgroup({key: ">"}, group_name, consumer) == {
+            key.encode(): {stream_id1_0.encode(): [[b"a", b"b"], [b"c", b"d"]]}
+        }
+
+        result = await glide_client.xinfo_stream(key)
+        assert result.get(b"length") == 1
+        expected_first_entry = [stream_id1_0.encode(), [b"a", b"b", b"c", b"d"]]
+        assert result.get(b"first-entry") == expected_first_entry
+
+        # only one entry exists, so first and last entry should be the same
+        assert result.get(b"last-entry") == expected_first_entry
+
+        # call XINFO STREAM with a byte string arg
+        result2 = await glide_client.xinfo_stream(key.encode())
+        assert result2 == result
+
+        # add one more entry
+        assert (
+            await glide_client.xadd(
+                key, [("foo", "bar")], StreamAddOptions(stream_id1_1)
+            )
+            == stream_id1_1.encode()
+        )
+
+        result_full = await glide_client.xinfo_stream_full(key, count=1)
+        print(result_full)
+        assert result_full.get(b"length") == 2
+        entries = cast(list, result_full.get(b"entries"))
+        # only the first entry will be returned since we passed count=1
+        assert len(entries) == 1
+        assert entries[0] == expected_first_entry
+
+        groups = cast(list, result_full.get(b"groups"))
+        assert len(groups) == 1
+        group_info = groups[0]
+        assert group_info.get(b"name") == group_name.encode()
+        pending = group_info.get(b"pending")
+        assert len(pending) == 1
+        assert stream_id1_0.encode() in pending[0]
+
+        consumers = group_info.get(b"consumers")
+        assert len(consumers) == 1
+        consumer_info = consumers[0]
+        assert consumer_info.get(b"name") == consumer.encode()
+        consumer_pending = consumer_info.get(b"pending")
+        assert len(consumer_pending) == 1
+        assert stream_id1_0.encode() in consumer_pending[0]
+
+        # call XINFO STREAM FULL with byte arg
+        result_full2 = await glide_client.xinfo_stream_full(key.encode())
+        # 2 entries should be returned, since we didn't pass the COUNT arg this time
+        assert len(cast(list, result_full2.get(b"entries"))) == 2
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_xinfo_stream_edge_cases_and_failures(
+        self, glide_client: TGlideClient, cluster_mode, protocol
+    ):
+        key = get_random_string(10)
+        string_key = get_random_string(10)
+        non_existing_key = get_random_string(10)
+        stream_id1_0 = "1-0"
+
+        # setup: create empty stream
+        assert (
+            await glide_client.xadd(
+                key, [("field", "value")], StreamAddOptions(stream_id1_0)
+            )
+            == stream_id1_0.encode()
+        )
+        assert await glide_client.xdel(key, [stream_id1_0]) == 1
+
+        # XINFO STREAM called against empty stream
+        result = await glide_client.xinfo_stream(key)
+        assert result.get(b"length") == 0
+        assert result.get(b"first-entry") is None
+        assert result.get(b"last-entry") is None
+
+        # XINFO STREAM FULL called against empty stream. Negative count values are ignored.
+        result_full = await glide_client.xinfo_stream_full(key, count=-3)
+        assert result_full.get(b"length") == 0
+        assert result_full.get(b"entries") == []
+        assert result_full.get(b"groups") == []
+
+        # calling XINFO STREAM with a non-existing key raises an error
+        with pytest.raises(RequestError):
+            await glide_client.xinfo_stream(non_existing_key)
+        with pytest.raises(RequestError):
+            await glide_client.xinfo_stream_full(non_existing_key)
+
+        # key exists, but it is not a stream
+        assert await glide_client.set(string_key, "foo")
+        with pytest.raises(RequestError):
+            await glide_client.xinfo_stream(string_key)
+        with pytest.raises(RequestError):
+            await glide_client.xinfo_stream_full(string_key)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_xgroup_set_id(
         self, glide_client: TGlideClient, cluster_mode, protocol, request
     ):
@@ -8234,6 +8353,125 @@ class TestCommands:
                 check_function_stats_response(
                     cast(TFunctionStatsResponse, node_response), [], 0, 0
                 )
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_function_kill_no_write(
+        self, request, cluster_mode, protocol, glide_client: TGlideClient
+    ):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        lib_name = f"mylib1C{get_random_string(5)}"
+        func_name = f"myfunc1c{get_random_string(5)}"
+        code = create_lua_lib_with_long_running_function(lib_name, func_name, 10, True)
+
+        # nothing to kill
+        with pytest.raises(RequestError) as e:
+            await glide_client.function_kill()
+        assert "NotBusy" in str(e)
+
+        # load the library
+        assert await glide_client.function_load(code, replace=True) == lib_name.encode()
+
+        # create a second client to run fcall
+        test_client = await create_client(
+            request, cluster_mode=cluster_mode, protocol=protocol, timeout=15000
+        )
+
+        # call fcall to run the function
+        # make sure that fcall routes to a primary node, and not a replica
+        # if this happens, function_kill and function_stats won't find the function and will fail
+        primaryRoute = SlotKeyRoute(SlotType.PRIMARY, lib_name)
+
+        async def endless_fcall_route_call():
+            # fcall is supposed to be killed, and will return a RequestError
+            with pytest.raises(RequestError) as e:
+                if cluster_mode:
+                    await test_client.fcall_ro_route(
+                        func_name, arguments=[], route=primaryRoute
+                    )
+                else:
+                    await test_client.fcall_ro(func_name, arguments=[])
+            assert "Script killed by user" in str(e)
+
+        async def wait_and_function_kill():
+            # it can take a few seconds for FCALL to register as running
+            await asyncio.sleep(3)
+            timeout = 0
+            while timeout <= 5:
+                # keep trying to kill until we get an "OK"
+                try:
+                    result = await glide_client.function_kill()
+                    #  we expect to get success
+                    assert result == "OK"
+                    break
+                except RequestError:
+                    # a RequestError may occur if the function is not yet running
+                    # sleep and try again
+                    timeout += 0.5
+                    await asyncio.sleep(0.5)
+
+        await asyncio.gather(
+            endless_fcall_route_call(),
+            wait_and_function_kill(),
+        )
+
+        # no functions running so we get notbusy error again
+        with pytest.raises(RequestError) as e:
+            assert await glide_client.function_kill()
+        assert "NotBusy" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [False, True])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_function_kill_write_is_unkillable(
+        self, request, cluster_mode, protocol, glide_client: TGlideClient
+    ):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        lib_name = f"mylib1C{get_random_string(5)}"
+        func_name = f"myfunc1c{get_random_string(5)}"
+        code = create_lua_lib_with_long_running_function(lib_name, func_name, 10, False)
+
+        # load the library on all primaries
+        assert await glide_client.function_load(code, replace=True) == lib_name.encode()
+
+        # create a second client to run fcall - and give it a long timeout
+        test_client = await create_client(
+            request, cluster_mode=cluster_mode, protocol=protocol, timeout=15000
+        )
+
+        # call fcall to run the function loaded function
+        async def endless_fcall_route_call():
+            # fcall won't be killed, because kill only works against fcalls that don't make a write operation
+            # use fcall(key) so that it makes a write operation
+            await test_client.fcall(func_name, keys=[lib_name])
+
+        async def wait_and_function_kill():
+            # it can take a few seconds for FCALL to register as running
+            await asyncio.sleep(3)
+            timeout = 0
+            foundUnkillable = False
+            while timeout <= 5:
+                # keep trying to kill until we get a unkillable return error
+                try:
+                    await glide_client.function_kill()
+                except RequestError as e:
+                    if "UNKILLABLE" in str(e):
+                        foundUnkillable = True
+                        break
+                timeout += 0.5
+                await asyncio.sleep(0.5)
+            # expect an unkillable error
+            assert foundUnkillable
+
+        await asyncio.gather(
+            endless_fcall_route_call(),
+            wait_and_function_kill(),
+        )
 
     @pytest.mark.parametrize("cluster_mode", [True])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])